requirements.txt
h5py
nltk
tqdm
pytorch
matplotlib